requirements.txt
cython
nltk
numpy==1.23
chumpy
requests
dv
opencv-python
matplotlib
imageio
imageio-ffmpeg
pyyaml
tqdm
boto3
scipy
yacs
roma
pandas
pyntcloud
tensorboard
einops